Only announces direct buffs that are targeted ON the player
Regular abilities such as Tranquility, Rallying Cry, Spirit Link Totem etc. is still announced.

destFlag 66833 is PLAYER (I could not find the correlating FINAL variable.

destFlag -2147483648 is abilities without a target (Integer.MIN) which we need to allow.

The Config change for saving the "onlySelf" variable is not as nice as I would like it, but it was the only way I could think of keeping it in the same group.